requirements.txt
Flask
flask-cors
boto3
python-magic
requests
gunicorn
Pillow
